+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Remove APEv2 and ID3v1 tags appended to mp3, ogg and flac files
 - Add avif support
 - Add jpeg xl (jxl) support
+- Stop reordering elements in office files, only sort their attributes
 
 # 0.14.0 - 2025-10-23
 - Add webp support